fix: change findByDiscussionReaction return type Optional<DiscussionReactionEntity> to nullable DiscussionReactionEntity

+++ src/main/kotlin/swm/virtuoso/reviewservice/adapter/out/persistence/DiscussionReactionPersistenceAdapter.kt
@@ -11,15 +11,17 @@ class DiscussionReactionPersistenceAdapter(
     private val discussionReactionRepository: DiscussionReactionRepository
 ) : DiscussionReactionPort {
     override fun insertReaction(discussionReaction: DiscussionReaction): DiscussionReaction {
-        val newReactionEntity = DiscussionReactionEntity.fromDiscussionReaction(discussionReaction)
-        val entity = discussionReactionRepository.save(newReactionEntity)
-        return DiscussionReaction.fromEntity(entity)
+        val reactionEntity = discussionReactionRepository.findByDiscussionReaction(discussionReaction)
+            ?: discussionReactionRepository.save(DiscussionReactionEntity.fromDiscussionReaction(discussionReaction))
+        return DiscussionReaction.fromEntity(reactionEntity)
     }
 
     override fun removeReaction(discussionReaction: DiscussionReaction): Boolean {
         val entity = discussionReactionRepository
             .findByDiscussionReaction(discussionReaction)
-            .orElseThrow { RuntimeException("Could not find DiscussionReactionEntity from DiscussionReaction") }
+        if (entity == null) {
+            throw RuntimeException("Could not find DiscussionReactionEntity from DiscussionReaction")
+        }
         return runCatching { discussionReactionRepository.delete(entity) }.isSuccess
     }
